// src/app/modules/campaign-setting/setting-point-incentive/models/PointToValueListModel.ts
export interface PointToValueListModel{
    pointToIncentiveRangeConfigurationId: number
    campaignSettingId?: number
    currencyId?: number
    rangeNo?: number
    incentiveValueAmount?: number
    validPointAmountFrom?: number
    validPointAmountTo?: number
    createdBy?: string
    createdDatetime?: string
    updatedBy?: string
    updatedDate?: string
}